LayoutTests/fast/css-grid-layout/grid-disabled-expected.txt

Verifies that display: -webkit-grid does nothing if runtime feature is disabled.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.getElementById('grid').style.display is ''
PASS successfullyParsed is true

TEST COMPLETE
